Get complete attachments, saved and unsaved (#103)
* Add getAttachments to PDFDocument

This adds the getAttachments method to get all the attachments from a
pdf document.

This is basically implementing the suggestion in
https://github.com/Hopding/pdf-lib/issues/534#issuecomment-662758789

There's some interesting behavior here, as the attachments don't show up
until after save is called. That's tested in the test and described in
the README, just so it's clear that's the behavior.

I don't know how to do all that great jsfiddle setup, so for the README
Ijust copied the basic example from that issue for now.

* remove duplicate identifier

This was caused by cherry-picking the get attachment branch.

* catch failed lookups

* catch failed lookups

* extract description, MIME type, and relationship

* test different attachment types

* get unsaved attachments

* also extract creation and modification date

* lint code

* remove obsolete hint

* expose embedder used by PDFEmbeddedFile

* remove test case

This is covered elsewhere.

---------

Co-authored-by: Shaun Verch <githubpublic@shaunverch.com>